+++ CHANGELOG.md
@@ -1,12 +1,29 @@
+## v0.2.1:
+
+date: 2017-01-20
+
+changes:
+
+- 重构 css 代码
+- 更新 API 文档
+- 修复 renderChapters 方法下获取 SUB_LIST 模版的 BUG（之前获取的模版是固定的，而不是用户配置的 SUB_LIST）
+- 移除原来 window 对象下的 autoc 方法
+- attachEvents 转为私有方法 _attachEvents
+- renderLinks 更名为 renderAnchors
+- 新增 isAnimateScroll 配置，实现配置锚点跳转或者动画滚动跳转定位
+- 新增 scrollTo 方法，实现页面动画滚动定位
+- 新增 onlyAnchors 配置，onlyAnchors 为 ture，则 AutocJS 的行为和 AnchorJS 一样。为 false， 则会出现导航菜单
+
+
 ## v0.2.0:
 
 date: 2017-01-19
 
 changes:
 
-- 重构代码
-- 更新API文档
-- 修复 getPidByDiffer 的BUG
+- 重构 js 代码
+- 更新 API 文档
+- 修复 getPidByDiffer 的 BUG
 
 
 ## v0.1.5:
@@ -34,6 +51,7 @@ changes:
 date: 2016-04-23
 
 changes:
+
   - 添加新的方法（renderAnchorLinks\setChapters\reload）
   - 致敬 AnchorJS，给 H1~H6 标签添加类似 AnchorJS 的链接。
   - 所有对外的方法都返回为单体对象，可以采用链式调用。
@@ -44,6 +62,7 @@ changes:
 date: 2016-04-08
 
 changes:
+
   - 重构 AutocJS，调整API结构，采用单体对象管理。
   - Repositories contributed to NPM & bower.
   - 修复 AMD 模块下文 BUG
@@ -54,6 +73,7 @@ changes:
 date: 2016-04-08
 
 changes:
+
   - 支持 AMD 和 CMD 规范.
   - 将 AutocJS 扩展成为一个 jQuery 插件.
 
@@ -63,4 +83,5 @@ changes:
 date: 2016-04-05
 
 changes:
+
   - AutocJS 程序的初始版本.